celeste/requirements.txt
mss
pillow
numpy
opencv-python
keyboard